+++ src/admin_ui/src/index.tsx
@@ -1,22 +1,17 @@
+import {createRoot} from 'react-dom/client';
 import React from 'react';
-import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
 import {AppParams} from "./admin/types";
 
-declare var __RESTFW_ADMIN_PARAMS__: AppParams;
+declare let __RESTFW_ADMIN_PARAMS__: AppParams;
 
-const root = ReactDOM.createRoot(
-    document.getElementById('root') as HTMLElement
-);
-root.render(
-    <React.StrictMode>
-        <App {...__RESTFW_ADMIN_PARAMS__}/>
-    </React.StrictMode>
-);
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
+const container = document.getElementById('root');
+if (container) {
+    const root = createRoot(container);
+    root.render(
+        <React.StrictMode>
+            <App {...__RESTFW_ADMIN_PARAMS__}/>
+        </React.StrictMode>
+    );
+}