+++ src/example_app/example_app/documents/__init__.py
@@ -3,6 +3,7 @@
 :Authors: cykooz
 :Date: 31.07.2020
 """
+
 from pyramid.config import Configurator
 
 
@@ -15,14 +16,20 @@ def includeme(config: Configurator):
 
     def add_to_root(event):
         docs = event.root['docs'] = Docs()
-        docs.http_post(event.request, params={
-            'user_id': 1,
-            'name': 'Important document',
-            'data': 'It is a document data.'
-        })
+        for i in range(1, 1001):
+            docs.http_post(
+                event.request,
+                params={
+                    'user_id': 1,
+                    'name': f'Important document {i}',
+                    'data': f'It is a data of document {i}.',
+                },
+            )
 
     from restfw.interfaces import IRootCreated
+
     config.add_subscriber(add_to_root, IRootCreated)
 
     from restfw.utils import scan_ignore
+
     config.scan(ignore=scan_ignore(config.registry))